requirements.txt
streamlit
openai-whisper
deep-translator
edge-tts
yt-dlp
pandas
requests
python-dotenv
